phase-ii/backend/requirements.txt
# Todo Full-Stack Web Application (Phase-2) - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database ORM
sqlmodel==0.0.14
asyncpg==0.29.0

# Database Migrations
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0.post1

# HTTP Client (for testing)
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
